+++ packages/engine-website-app/build.config.js
@@ -1,3 +1,4 @@
+/* eslint-disable no-undef */
 import { viteStaticCopy } from 'vite-plugin-static-copy';
 import pkg from './package.json';
 import commonConfig from './build.common.config';
@@ -31,6 +32,7 @@ const mainConfig = {
     plugins: plugins,
     ...commonConfig,
     define: {
+      global: 'globalThis',
       'process.env.NODE_ENV': JSON.stringify(process.env.NODE_ENV),
       __RUN_MODE__: JSON.stringify(process.env.BUILD_TYPE),
       __PACKAGE_VERSION__: JSON.stringify(pkg.version),