+++ packages/engine-website-app/src/page/Preview/index.tsx
@@ -63,7 +63,7 @@ export const Preview = () => {
         }}
         render={renderHandle}
         adapter={ReactAdapter}
-        $$context={renderContext}
+        $$context={renderContext as any}
       />
     </div>
   );